requirements.txt
beautifulsoup4==4.10.0
requests2==2.16.0
